application/model_training/requirements.txt
torch
torchvision
numpy
scikit-learn
tqdm
Pillow
